@@ -2,12 +2,12 @@ import { INCREMENT_COUNTER, DECREMENT_COUNTER } from '../constants';
 
 export function increment() {
   return {
-    type: INCREMENT_COUNTER,
+    type: INCREMENT_COUNTER
   };
 }
 
 export function decrement() {
   return {
-    type: DECREMENT_COUNTER,
+    type: DECREMENT_COUNTER
   };
 }
