@@ -7,9 +7,9 @@ export function post(path, data) {
     method: 'post',
     headers: {
       'Accept': 'application/json',
-      'Content-Type': 'application/json',
+      'Content-Type': 'application/json'
     },
-    body: JSON.stringify(data),
+    body: JSON.stringify(data)
   })
   .then(response => response.json());
 }
