@@ -10,6 +10,6 @@
 export default function fireAction(reducer, currentState, type, payload = {}) {
   return reducer(currentState, {
     type,
-    payload,
+    payload
   });
 }
