@@ -3,12 +3,12 @@ const production = process.env.NODE_ENV === 'production';
 
 exports.eslint = {
   test: /\.jsx?$/,
-  loader: 'eslint-loader',
+  loader: 'eslint-loader'
 };
 
 exports.css = {
   test: /\.css$/,
-  loader: 'style-loader!css?-minimize!postcss',
+  loader: 'style-loader!css?-minimize!postcss'
 };
 
 const jsloaders = [];
@@ -24,29 +24,29 @@ if (testing) {
 exports.js = {
   test: /\.jsx?$/,
   loaders: jsloaders,
-  exclude: /node_modules/,
+  exclude: /node_modules/
 };
 
 exports.istanbulInstrumenter = {
   test: /^(.(?!test))*\.jsx?$/,
   loader: 'istanbul-instrumenter-loader',
   query: {
-    embedSource: true,
+    embedSource: true
   },
-  exclude: /node_modules/,
+  exclude: /node_modules/
 };
 
 exports.json = {
   test: /\.json$/,
-  loader: 'json-loader',
+  loader: 'json-loader'
 };
 
 exports.image = {
   test: /\.(png|jpg|jpeg|gif|svg)$/,
-  loader: 'url-loader?prefix=img/&limit=5000',
+  loader: 'url-loader?prefix=img/&limit=5000'
 };
 
 exports.font = {
   test: /\.(woff|woff2|ttf|eot)$/,
-  loader: 'url-loader?prefix=font/&limit=5000',
+  loader: 'url-loader?prefix=font/&limit=5000'
 };
