@@ -9,36 +9,36 @@ const basePlugins = [
   new webpack.DefinePlugin({
     __DEV__: process.env.NODE_ENV !== 'production',
     __TEST__: JSON.stringify(process.env.TEST),
-    'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV),
+    'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV)
   }),
   new HtmlWebpackPlugin({
     template: './src/index.html',
-    inject: 'body',
+    inject: 'body'
   }),
   new webpack.NoErrorsPlugin(),
   new CopyWebpackPlugin([
-    { from: 'src/assets', to: 'assets' },
-  ]),
+    { from: 'src/assets', to: 'assets' }
+  ])
 ];
 
 const devPlugins = [
   new StyleLintPlugin({
     configFile: './.stylelintrc',
     files: ['src/**/*.css'],
-    failOnError: false,
-  }),
+    failOnError: false
+  })
 ];
 
 const prodPlugins = [
   new SplitByPathPlugin([
-    { name: 'vendor', path: [path.join(__dirname, '..', 'node_modules/')] },
+    { name: 'vendor', path: [path.join(__dirname, '..', 'node_modules/')] }
   ]),
   new webpack.optimize.UglifyJsPlugin({
     sourceMap: true,
     compress: {
-      warnings: false,
-    },
-  }),
+      warnings: false
+    }
+  })
 ];
 
 module.exports = basePlugins
